fix: move app test helpers to new pkg
